+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/custom/condition/NovelPagingCondition.java
@@ -1,0 +1,9 @@
+package com.iucyh.novelservice.novel.repository.custom.condition;
+
+import com.iucyh.novelservice.novel.repository.custom.paging.cursor.NovelCursor;
+
+public record NovelPagingCondition(
+
+        NovelCursor cursor,
+        int limit
+) {}
