+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/custom/paging/NovelLastUpdatePagingStrategy.java
@@ -1,10 +1,10 @@
-package com.iucyh.novelservice.novel.repository.query.paging;
+package com.iucyh.novelservice.novel.repository.custom.paging;
 
 import com.iucyh.novelservice.novel.domain.Novel;
-import com.iucyh.novelservice.novel.repository.query.paging.cursor.NovelCursor;
-import com.iucyh.novelservice.novel.repository.query.paging.cursor.NovelPopularCursor;
+import com.iucyh.novelservice.novel.repository.custom.paging.cursor.NovelCursor;
+import com.iucyh.novelservice.novel.repository.custom.paging.cursor.NovelLastUpdateCursor;
 import com.iucyh.novelservice.novel.enumtype.NovelSortType;
-import com.iucyh.novelservice.novel.repository.query.paging.template.AbstractNovelPagingStrategy;
+import com.iucyh.novelservice.novel.repository.custom.paging.template.AbstractNovelPagingStrategy;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import org.springframework.stereotype.Component;
@@ -12,34 +12,34 @@ import org.springframework.stereotype.Component;
 import static com.iucyh.novelservice.novel.domain.QNovel.novel;
 
 @Component
-public class NovelPopularPagingStrategy extends AbstractNovelPagingStrategy {
+public class NovelLastUpdatePagingStrategy extends AbstractNovelPagingStrategy {
 
     @Override
     protected OrderSpecifier<?>[] applyOrder() {
         return new OrderSpecifier[] {
-                novel.periodViewCount.desc(),
                 novel.lastEpisodePublishDate.desc(),
+                novel.totalViewCount.desc(),
                 novel.id.desc()
         };
     }
 
     @Override
     protected BooleanExpression applyCursor(NovelCursor cursor) {
-        NovelPopularCursor popularCursor = (NovelPopularCursor) cursor;
-        return novel.periodViewCount.lt(popularCursor.getPeriodViewCount())
+        NovelLastUpdateCursor lastUpdateCursor = (NovelLastUpdateCursor) cursor;
+        return novel.lastEpisodePublishDate.lt(lastUpdateCursor.getLastEpisodePublishDate())
                 .or(
-                        novel.periodViewCount.eq(popularCursor.getPeriodViewCount())
+                        novel.lastEpisodePublishDate.eq(lastUpdateCursor.getLastEpisodePublishDate())
                                 .and(
-                                        novel.lastEpisodePublishDate.lt(popularCursor.getLastEpisodePublishDate())
+                                        novel.totalViewCount.lt(lastUpdateCursor.getTotalViewCount())
                                 )
                 )
                 .or(
-                        novel.periodViewCount.eq(popularCursor.getPeriodViewCount())
+                        novel.lastEpisodePublishDate.eq(lastUpdateCursor.getLastEpisodePublishDate())
                                 .and(
-                                        novel.lastEpisodePublishDate.eq(popularCursor.getLastEpisodePublishDate())
+                                        novel.totalViewCount.eq(lastUpdateCursor.getTotalViewCount())
                                 )
                                 .and(
-                                        novel.id.lt(popularCursor.getNovelId())
+                                        novel.id.lt(lastUpdateCursor.getNovelId())
                                 )
                 );
     }
@@ -52,16 +52,16 @@ public class NovelPopularPagingStrategy extends AbstractNovelPagingStrategy {
 
     @Override
     public NovelCursor createCursor(Novel lastResult) {
-        return NovelPopularCursor.of(lastResult.getId(), lastResult.getPeriodViewCount(), lastResult.getLastEpisodePublishDate());
+        return NovelLastUpdateCursor.of(lastResult.getId(), lastResult.getLastEpisodePublishDate(), lastResult.getTotalViewCount());
     }
 
     @Override
     public Class<? extends NovelCursor> getCursorClass() {
-        return NovelPopularCursor.class;
+        return NovelLastUpdateCursor.class;
     }
 
     @Override
     public NovelSortType getSupportedSortType() {
-        return NovelSortType.POPULAR;
+        return NovelSortType.LAST_UPDATE;
     }
 }
