+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/custom/paging/NovelPagingStrategy.java
@@ -1,7 +1,7 @@
-package com.iucyh.novelservice.novel.repository.query.paging;
+package com.iucyh.novelservice.novel.repository.custom.paging;
 
 import com.iucyh.novelservice.novel.domain.Novel;
-import com.iucyh.novelservice.novel.repository.query.paging.cursor.NovelCursor;
+import com.iucyh.novelservice.novel.repository.custom.paging.cursor.NovelCursor;
 import com.iucyh.novelservice.novel.enumtype.NovelSortType;
 import com.querydsl.jpa.impl.JPAQuery;
