--- novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/condition/NovelPagingCondition.java
@@ -1,9 +1,0 @@
-package com.iucyh.novelservice.novel.repository.query.condition;
-
-import com.iucyh.novelservice.novel.repository.query.paging.cursor.NovelCursor;
-
-public record NovelPagingCondition(
-
-        NovelCursor cursor,
-        int limit
-) {}
